src/mighty_gobbla/backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
pdf2image
aiofiles
pypdf
